Добавил папку DAL, а в нее класс для контекста базы данных

+++ Diploma/Models/User.cs
@@ -8,7 +8,7 @@ namespace Diploma.Models
 {
     class User
     {
-        public long userId { get; set; }
+        public int userId { get; set; }
         public string login { get; set; }
         public string password { get; set; }
         public virtual ICollection<UserData> UserDatas { get; set; }
